+++ CHANGELOG.md
@@ -30,6 +30,7 @@ Fixed:
 Added:
 
 - Niche cluster connectivity visualization for each sample
+- `scale factor` controlling the size of spatial-based plots
 
 Changed: